postgres data type int to real

+++ postgres_dialect.go
@@ -8,7 +8,7 @@ func (p PostgresDialect) CreateTableSQL() string {
 	return `CREATE TABLE IF NOT EXISTS db_version
                 (
                     id             SERIAL                  NOT NULL,
-                    version        INTEGER                 NOT NULL,
+                    version        REAL	                   NOT NULL,
                     description    CHARACTER VARYING (255) NOT NULL,
                     checksum       CHARACTER VARYING (32)  NOT NULL,
                     applied_at     INTEGER                 NOT NULL,
